backend/requirements.txt
flask
amadeus
psycopg2-binary
simplejson
pytest
requests
dotenv
csv